chore: rename RegularButton files into Button
